// src/routes/names.enum.ts
export enum RoutesName {
  HOUSESELECT = 'HouseScreen',
  PROFILE = 'Profile',
  CHAT = 'ChatScreen',
  REPORT_HOME = 'ReportHome',
  REPORT = 'Report',
  LIST_PROYECTS = 'ListProyectsScreen',
  LIST_PROPERTIES = 'ListProperties',
  PROFILE_OPERATOR = 'ProfileOperator',
  TICKET_OPERATOR = 'TicketOperator',
}